requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
line-bot-sdk==1.16.0
line-pay==0.1.0
google-cloud-datastore==1.11.0
# gunicorn
# Keras
# absl-py
# grpcio
# Markdown
# protobuf
# Werkzeug
# tensorflow
# Pillow